Move all go code under pkg/
