+++ pkg/networkservicemesh/core/adapters/server_to_client.go
