+++ pkg/networkservicemesh/core/chain/client.go
@@ -18,8 +18,8 @@ package chain
 import (
 	"github.com/networkservicemesh/networkservicemesh/controlplane/api/networkservice"
 
-	"github.com/networkservicemesh/sdk/networkservicemesh/core/next"
-	"github.com/networkservicemesh/sdk/networkservicemesh/core/trace"
+	"github.com/networkservicemesh/sdk/pkg/networkservicemesh/core/next"
+	"github.com/networkservicemesh/sdk/pkg/networkservicemesh/core/trace"
 )
 
 func NewNetworkServiceClient(clients ...networkservice.NetworkServiceClient) networkservice.NetworkServiceClient {
