+++ pkg/networkservicemesh/core/trace/context.go
