+++ pkg/networkservicemesh/core/trace/server.go
@@ -21,24 +21,23 @@ import (
 	"fmt"
 
 	"github.com/golang/protobuf/ptypes/empty"
-	"google.golang.org/grpc"
 
 	"github.com/networkservicemesh/networkservicemesh/controlplane/api/connection"
 	"github.com/networkservicemesh/networkservicemesh/controlplane/api/networkservice"
 
-	"github.com/networkservicemesh/sdk/tools/spanhelper"
-	"github.com/networkservicemesh/sdk/tools/typeutils"
+	"github.com/networkservicemesh/sdk/pkg/tools/spanhelper"
+	"github.com/networkservicemesh/sdk/pkg/tools/typeutils"
 )
 
-type traceClient struct {
-	traced networkservice.NetworkServiceClient
+type traceServer struct {
+	traced networkservice.NetworkServiceServer
 }
 
-func NewNetworkServiceClient(traced networkservice.NetworkServiceClient) networkservice.NetworkServiceClient {
-	return &traceClient{traced: traced}
+func NewNetworkServiceServer(traced networkservice.NetworkServiceServer) networkservice.NetworkServiceServer {
+	return &traceServer{traced: traced}
 }
 
-func (t *traceClient) Request(ctx context.Context, request *networkservice.NetworkServiceRequest, opts ...grpc.CallOption) (*connection.Connection, error) {
+func (t *traceServer) Request(ctx context.Context, request *networkservice.NetworkServiceRequest) (*connection.Connection, error) {
 	// Create a new span
 	span := spanhelper.FromContext(ctx, fmt.Sprintf("%s.Request", typeutils.GetTypeName(t.traced)))
 	defer span.Finish()
@@ -50,7 +49,7 @@ func (t *traceClient) Request(ctx context.Context, request *networkservice.Netwo
 	span.LogObject("request", request)
 
 	// Actually call the next
-	rv, err := t.traced.Request(ctx, request, opts...)
+	rv, err := t.traced.Request(ctx, request)
 
 	if err != nil {
 		span.LogError(err)
@@ -60,7 +59,7 @@ func (t *traceClient) Request(ctx context.Context, request *networkservice.Netwo
 	return rv, err
 }
 
-func (t *traceClient) Close(ctx context.Context, conn *connection.Connection, opts ...grpc.CallOption) (*empty.Empty, error) {
+func (t *traceServer) Close(ctx context.Context, conn *connection.Connection) (*empty.Empty, error) {
 	// Create a new span
 	span := spanhelper.FromContext(ctx, fmt.Sprintf("%s.Close", typeutils.GetTypeName(t.traced)))
 	defer span.Finish()
@@ -68,7 +67,7 @@ func (t *traceClient) Close(ctx context.Context, conn *connection.Connection, op
 	ctx = withLog(span.Context(), span.Logger())
 
 	span.LogObject("request", conn)
-	rv, err := t.traced.Close(ctx, conn, opts...)
+	rv, err := t.traced.Close(ctx, conn)
 
 	if err != nil {
 		span.LogError(err)
